+++ Drone/SharpSploit/Evasion/Etw.cs
@@ -1,22 +1,21 @@
 ﻿using System;
-using System.Data.SqlTypes;
 using System.Runtime.InteropServices;
 
 namespace Drone.SharpSploit.Evasion
 {
-    public static class Etw
-    {
-	    public static EtwEventWriteDelegate EtwEventWriteOriginal { get; set; }
-	    
-        [UnmanagedFunctionPointer(CallingConvention.StdCall, CharSet=CharSet.Auto)]
-        public delegate ulong EtwEventWriteDelegate(
-            IntPtr RegHandle,
-            EVENT_DESCRIPTOR EventDescriptor,
-            ulong UserDataCount,
-            EVENT_DATA_DESCRIPTOR UserData);
-        
-        [StructLayout(LayoutKind.Sequential)]
-        public struct EVENT_DESCRIPTOR
+	public static class Etw
+	{
+		public static EtwEventWriteDelegate EtwEventWriteOriginal { get; set; }
+
+		[UnmanagedFunctionPointer(CallingConvention.StdCall, CharSet = CharSet.Auto)]
+		public delegate ulong EtwEventWriteDelegate(
+			IntPtr RegHandle,
+			EVENT_DESCRIPTOR EventDescriptor,
+			ulong UserDataCount,
+			EVENT_DATA_DESCRIPTOR UserData);
+
+		[StructLayout(LayoutKind.Sequential)]
+		public struct EVENT_DESCRIPTOR
 		{
 			public ushort Id;
 			public byte Version;
@@ -26,18 +25,18 @@ namespace Drone.SharpSploit.Evasion
 			public ushort Task;
 			public ulong Keyword;
 		}
-        
-        [StructLayout(LayoutKind.Sequential)]
-        public struct EVENT_DATA_DESCRIPTOR
-        {
-	        ulong Ptr;
-	        ulong Size;
-	        ulong Reserved;
-        }
 
-        public static ulong EtwEventWriteDetour(IntPtr RegHandle, EVENT_DESCRIPTOR EventDescriptor, ulong UserDataCount, EVENT_DATA_DESCRIPTOR UserData)
-        {
-	        return 0;
-        } 
-    }
+		[StructLayout(LayoutKind.Sequential)]
+		public struct EVENT_DATA_DESCRIPTOR
+		{
+			ulong Ptr;
+			ulong Size;
+			ulong Reserved;
+		}
+
+		public static ulong EtwEventWriteDetour(IntPtr RegHandle, EVENT_DESCRIPTOR EventDescriptor, ulong UserDataCount, EVENT_DATA_DESCRIPTOR UserData)
+		{
+			return 0;
+		}
+	}
 }